+++ src/CoinbaseSdk/Prime/portfolios/PortfoliosService.cs
@@ -47,11 +47,11 @@ namespace CoinbaseSdk.Prime.Portfolios
         cancellationToken);
     }
 
-    public GetPortfolioByIdResponse GetPortfolioById(
-      GetPortfolioByIdRequest request,
+    public GetPortfolioResponse GetPortfolio(
+      GetPortfolioRequest request,
       CallOptions? options = null)
     {
-      return this.Request<GetPortfolioByIdResponse>(
+      return this.Request<GetPortfolioResponse>(
         HttpMethod.Get,
         $"/portfolios/{request.PortfolioId}",
         [HttpStatusCode.OK],
@@ -59,12 +59,12 @@ namespace CoinbaseSdk.Prime.Portfolios
         options);
     }
 
-    public Task<GetPortfolioByIdResponse> GetPortfolioByIdAsync(
-      GetPortfolioByIdRequest request,
+    public Task<GetPortfolioResponse> GetPortfolioAsync(
+      GetPortfolioRequest request,
       CallOptions? options = null,
       CancellationToken cancellationToken = default)
     {
-      return this.RequestAsync<GetPortfolioByIdResponse>(
+      return this.RequestAsync<GetPortfolioResponse>(
         HttpMethod.Get,
         $"/portfolios/{request.PortfolioId}",
         [HttpStatusCode.OK],