+++ src/CoinbaseSdk/Prime/transactions/GetTransactionRequest.cs
@@ -14,23 +14,33 @@
  * limitations under the License.
  */
 
-namespace CoinbaseSdk.Prime.Portfolios
+namespace CoinbaseSdk.Prime.Transactions
 {
   using System.Text.Json.Serialization;
   using CoinbaseSdk.Core.Error;
 
-  public class GetPortfolioByIdRequest(string portfolioId)
+  public class GetTransactionRequest(string portfolioId, string transactionId)
   {
     [JsonIgnore]
     public string PortfolioId { get; set; } = portfolioId;
 
-    public class GetPortfolioByIdRequestBuilder
+    [JsonIgnore]
+    public string TransactionId { get; set; } = transactionId;
+
+    public class GetTransactionRequestBuilder
     {
       private string? _portfolioId;
+      private string? _transactionId;
 
-      public GetPortfolioByIdRequestBuilder WithPortfolioId(string portfolioId)
+      public GetTransactionRequestBuilder WithPortfolioId(string portfolioId)
       {
         this._portfolioId = portfolioId;
+        return this;
+      }
+
+      public GetTransactionRequestBuilder WithTransactionId(string transactionId)
+      {
+        this._transactionId = transactionId;
         return this;
       }
 
@@ -38,24 +48,29 @@ namespace CoinbaseSdk.Prime.Portfolios
       /// Validate the builder.
       /// </summary>
       /// <exception cref="CoinbaseClientException">Thrown when the
-      /// <see cref="_portfolioId"/> is null, empty or whitespace.</exception>
+      /// <see cref="_portfolioId"/> or <see cref="_transactionId"/> are null, empty
+      /// or whitespace.</exception>
       private void Validate()
       {
         if (string.IsNullOrWhiteSpace(this._portfolioId))
         {
-          throw new CoinbaseClientException("PortfolioId cannot be null or empty");
+          throw new CoinbaseClientException("PortfolioId is required");
+        }
+        if (string.IsNullOrWhiteSpace(this._transactionId))
+        {
+          throw new CoinbaseClientException("TransactionId is required");
         }
       }
 
       /// <summary>
-      /// Build the <see cref="GetPortfolioByIdRequest"/>.
+      /// Build the <see cref="GetTransactionRequest"/> object.
       /// </summary>
-      /// <returns>The <see cref="GetPortfolioByIdRequest"/>.</returns>
+      /// <returns>The <see cref="GetTransactionRequest"/> object.</returns>
       /// <exception cref="CoinbaseClientException">Thrown when the required fields are not set.</exception>
-      public GetPortfolioByIdRequest Build()
+      public GetTransactionRequest Build()
       {
         this.Validate();
-        return new GetPortfolioByIdRequest(this._portfolioId!);
+        return new GetTransactionRequest(_portfolioId!, _transactionId!);
       }
     }
   }